refactor(domain/path): path 패키지 생성
