+++ src/main/java/wooteco/subway/service/PathService.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import wooteco.subway.domain.Fare;
-import wooteco.subway.domain.Path;
+import wooteco.subway.domain.path.Path;
 import wooteco.subway.domain.line.Line;
 import wooteco.subway.domain.station.Stations;
 import wooteco.subway.dto.PathResponse;
